CB-42 refactor: created app package to move helper module too
